+++ CHANGELOG.md
@@ -7,3 +7,4 @@
   - Handle both single and double quoted @import statements [#168](https://github.com/adamrenklint/asimov.js/issues/168)
   - Simple redirects with page.attributes.alias [#45](https://github.com/adamrenklint/asimov.js/issues/45)
   - Automatically reload page when build folder changes [#62](https://github.com/adamrenklint/asimov.js/issues/62)
+  - Fixed menu helper not applying active class to active page [#142](https://github.com/adamrenklint/asimov.js/issues/142)